+++ clash_lib/src/proxy/shadowsocks/mod.rs
@@ -20,7 +20,6 @@ use crate::{
     impl_default_connector,
     proxy::{CommonOption, OutboundHandler},
     session::Session,
-    Error,
 };
 use async_trait::async_trait;
 use shadowsocks::{
@@ -42,35 +41,6 @@ pub struct SimpleOBFSOption {
     pub host: String,
 }
 
-impl TryFrom<HashMap<String, serde_yaml::Value>> for SimpleOBFSOption {
-    type Error = crate::Error;
-
-    fn try_from(
-        value: HashMap<String, serde_yaml::Value>,
-    ) -> Result<Self, Self::Error> {
-        let host = value
-            .get("host")
-            .and_then(|x| x.as_str())
-            .unwrap_or("bing.com");
-        let mode = value
-            .get("mode")
-            .and_then(|x| x.as_str())
-            .ok_or(Error::InvalidConfig("obfs mode is required".to_owned()))?;
-
-        match mode {
-            "http" => Ok(SimpleOBFSOption {
-                mode: SimpleOBFSMode::Http,
-                host: host.to_owned(),
-            }),
-            "tls" => Ok(SimpleOBFSOption {
-                mode: SimpleOBFSMode::Tls,
-                host: host.to_owned(),
-            }),
-            _ => Err(Error::InvalidConfig(format!("invalid obfs mode: {}", mode))),
-        }
-    }
-}
-
 #[allow(dead_code)]
 pub struct V2RayOBFSOption {
     pub mode: String,
@@ -82,94 +52,11 @@ pub struct V2RayOBFSOption {
     pub mux: bool,
 }
 
-impl TryFrom<HashMap<String, serde_yaml::Value>> for V2RayOBFSOption {
-    type Error = crate::Error;
-
-    fn try_from(
-        value: HashMap<String, serde_yaml::Value>,
-    ) -> Result<Self, Self::Error> {
-        let host = value
-            .get("host")
-            .and_then(|x| x.as_str())
-            .unwrap_or("bing.com");
-        let mode = value
-            .get("mode")
-            .and_then(|x| x.as_str())
-            .ok_or(Error::InvalidConfig("obfs mode is required".to_owned()))?;
-
-        if mode != "websocket" {
-            return Err(Error::InvalidConfig(format!(
-                "invalid obfs mode: {}",
-                mode
-            )));
-        }
-
-        let path = value
-            .get("path")
-            .and_then(|x| x.as_str())
-            .ok_or(Error::InvalidConfig("obfs path is required".to_owned()))?;
-        let mux = value.get("mux").and_then(|x| x.as_bool()).unwrap_or(false);
-        let tls = value.get("tls").and_then(|x| x.as_bool()).unwrap_or(false);
-        let skip_cert_verify = value
-            .get("skip-cert-verify")
-            .and_then(|x| x.as_bool())
-            .unwrap_or(false);
-
-        let mut headers = HashMap::new();
-        if let Some(h) = value.get("headers") {
-            if let Some(h) = h.as_mapping() {
-                for (k, v) in h {
-                    if let (Some(k), Some(v)) = (k.as_str(), v.as_str()) {
-                        headers.insert(k.to_owned(), v.to_owned());
-                    }
-                }
-            }
-        }
-
-        Ok(V2RayOBFSOption {
-            mode: mode.to_owned(),
-            host: host.to_owned(),
-            path: path.to_owned(),
-            tls,
-            headers,
-            skip_cert_verify,
-            mux,
-        })
-    }
-}
-
 #[derive(Debug)]
 pub struct ShadowTlsOption {
     pub host: String,
     pub password: String,
     pub strict: bool,
-}
-
-impl TryFrom<HashMap<String, serde_yaml::Value>> for ShadowTlsOption {
-    type Error = crate::Error;
-
-    fn try_from(
-        value: HashMap<String, serde_yaml::Value>,
-    ) -> Result<Self, Self::Error> {
-        let host = value
-            .get("host")
-            .and_then(|x| x.as_str())
-            .unwrap_or("bing.com");
-        let password = value
-            .get("password")
-            .and_then(|x| x.as_str().to_owned())
-            .ok_or(Error::InvalidConfig("obfs mode is required".to_owned()))?;
-        let strict = value
-            .get("strict")
-            .and_then(|x| x.as_bool())
-            .unwrap_or(true);
-
-        Ok(Self {
-            host: host.to_string(),
-            password: password.to_string(),
-            strict,
-        })
-    }
 }
 
 pub enum OBFSOption {
